applied cr
