+++ src/sentry/static/sentry/app/components/events/interfaces/debugMeta-v2/utils.tsx
@@ -1,4 +1,5 @@
-import {ImageStatus} from 'app/types/debugImage';
+import {Image, ImageStatus} from 'app/types/debugImage';
+import {defined} from 'app/utils';
 
 export const IMAGE_AND_CANDIDATE_LIST_MAX_HEIGHT = 400;
 
@@ -36,4 +37,26 @@ export function getFileName(path?: string | null) {
 
 export function normalizeId(id?: string) {
   return id?.trim().toLowerCase().replace(/[- ]/g, '') ?? '';
+}
+
+// TODO(ts): When replacing debugMeta with debugMetaV2, also replace {type: string} with the Image type defined in 'app/types/debugImage'
+export function shouldSkipSection(
+  filteredImages: Array<{type: string}>,
+  images: Array<{type: string} | null>
+) {
+  if (!!filteredImages.length) {
+    return false;
+  }
+
+  const definedImages = images.filter(image => defined(image));
+
+  if (!definedImages.length) {
+    return true;
+  }
+
+  if ((definedImages as Array<Image>).every(image => image.type === 'proguard')) {
+    return true;
+  }
+
+  return false;
 }
