+++ src/sentry/static/sentry/app/components/events/interfaces/debugMeta/index.tsx
@@ -26,6 +26,8 @@ import {Frame, Organization, Project} from 'app/types';
 import {Event} from 'app/types/event';
 import EmptyMessage from 'app/views/settings/components/emptyMessage';
 
+import {shouldSkipSection} from '../debugMeta-v2/utils';
+
 import DebugImage from './debugImage';
 import {getFileName} from './utils';
 
@@ -192,6 +194,11 @@ class DebugMeta extends React.PureComponent<Props, State> {
     const foundFrame = this.getFrame();
     // skip null values indicating invalid debug images
     const debugImages = this.getDebugImages();
+
+    if (!debugImages.length) {
+      return;
+    }
+
     const filteredImages = debugImages.filter(image => this.filterImage(image));
 
     this.setState({debugImages, filteredImages, foundFrame});
@@ -406,6 +413,12 @@ class DebugMeta extends React.PureComponent<Props, State> {
 
   render() {
     const {filteredImages, foundFrame} = this.state;
+    const {data} = this.props;
+    const {images} = data;
+
+    if (shouldSkipSection(filteredImages, images)) {
+      return null;
+    }
 
     return (
       <StyledEventDataSection
